+++ Blahker/AppDelegate.swift
@@ -8,8 +8,6 @@
 
 import UIKit
 import SafariServices
-import Fabric
-import Crashlytics
 
 let contentBlockerExteiosnIdentifier = "com.elaborapp.Blahker.ContentBlocker"
 
@@ -22,9 +20,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         // Override point for customization after application launch.
 
         application.setMinimumBackgroundFetchInterval(UIApplicationBackgroundFetchIntervalMinimum)
-
-        //Crashlytics
-        Fabric.with([Crashlytics.self])
 
         return true
     }